added checks to migrations if ecommerceframework is enabled.

+++ bundles/CoreBundle/Resources/migrations/Version20191125200416.php
@@ -5,6 +5,7 @@ namespace Pimcore\Bundle\CoreBundle\Migrations;
 use Doctrine\DBAL\Schema\Schema;
 use Pimcore\Bundle\EcommerceFrameworkBundle\Factory;
 use Pimcore\Bundle\EcommerceFrameworkBundle\IndexService\Worker\AbstractBatchProcessingWorker;
+use Pimcore\Bundle\EcommerceFrameworkBundle\PimcoreEcommerceFrameworkBundle;
 use Pimcore\Migrations\Migration\AbstractPimcoreMigration;
 
 /**
@@ -17,25 +18,27 @@ class Version20191125200416 extends AbstractPimcoreMigration
      */
     public function up(Schema $schema)
     {
-        $factory = Factory::getInstance();
-        $indexService = $factory->getIndexService();
-        $tenants = $indexService->getTenants();
+        if (PimcoreEcommerceFrameworkBundle::isEnabled()) {
+            $factory = Factory::getInstance();
+            $indexService = $factory->getIndexService();
+            $tenants = $indexService->getTenants();
 
-        foreach($tenants as $tenant) {
+            foreach($tenants as $tenant) {
 
-            $tenantWorker = $indexService->getTenantWorker($tenant);
-            if($tenantWorker instanceof AbstractBatchProcessingWorker) {
+                $tenantWorker = $indexService->getTenantWorker($tenant);
+                if($tenantWorker instanceof AbstractBatchProcessingWorker) {
 
-                $method = new \ReflectionMethod(get_class($tenantWorker), 'getStoreTableName');
-                $method->setAccessible(true);
-                $tableName = $method->invoke($tenantWorker);
+                    $method = new \ReflectionMethod(get_class($tenantWorker), 'getStoreTableName');
+                    $method->setAccessible(true);
+                    $tableName = $method->invoke($tenantWorker);
 
-                $this->addSql("ALTER TABLE `$tableName`
+                    $this->addSql("ALTER TABLE `$tableName`
                     ADD metadata TEXT NULL;
                 ");
 
-            }
+                }
 
+            }
         }
 
     }
@@ -46,27 +49,28 @@ class Version20191125200416 extends AbstractPimcoreMigration
     public function down(Schema $schema)
     {
 
-        $factory = Factory::getInstance();
-        $indexService = $factory->getIndexService();
-        $tenants = $indexService->getTenants();
+        if (PimcoreEcommerceFrameworkBundle::isEnabled()) {
+            $factory = Factory::getInstance();
+            $indexService = $factory->getIndexService();
+            $tenants = $indexService->getTenants();
 
-        foreach($tenants as $tenant) {
+            foreach ($tenants as $tenant) {
 
-            $tenantWorker = $indexService->getTenantWorker($tenant);
-            if($tenantWorker instanceof AbstractBatchProcessingWorker) {
+                $tenantWorker = $indexService->getTenantWorker($tenant);
+                if ($tenantWorker instanceof AbstractBatchProcessingWorker) {
 
-                $method = new \ReflectionMethod(get_class($tenantWorker), 'getStoreTableName');
-                $method->setAccessible(true);
-                $tableName = $method->invoke($tenantWorker);
+                    $method = new \ReflectionMethod(get_class($tenantWorker), 'getStoreTableName');
+                    $method->setAccessible(true);
+                    $tableName = $method->invoke($tenantWorker);
 
-                $this->addSql("ALTER TABLE `$tableName`
+                    $this->addSql("ALTER TABLE `$tableName`
                     DROP COLUMN metadata;
                 ");
 
+                }
+
             }
-
         }
-
 
     }
 }
